added fixed sql tables

+++ CoreBankingSystem.DAL/Models/Transaction.cs
@@ -8,8 +8,8 @@ namespace CoreBankingSystem.DAL.Models
         public DateTime TransactionDate { get; set; }
         public decimal Amount { get; set; }
         public TransactionType TransactionType { get; set; }
-        public int SenderAccount { get; set; }
-        public int ReceiverAccount { get; set; }
+        public long SenderAccount { get; set; }
+        public long ReceiverAccount { get; set; }
 
         public int AccountId { get; set; }
         public Account Account { get; set; }
